rename KompetanseContext -> useKompetanse og KompetanseSkjemaContext -> useKompetansePeriodeSkjema
