+++ src/frontend/context/Kompetanse/useKompetanse.ts
